requirements.txt
python-telegram-bot==20.6
flask
pandas
requests
gunicorn